requirements.txt
# Core dependencies
textual>=0.40.0
rich>=13.0.0
stem>=1.8.0
PySocks>=1.7.1
cryptography>=41.0.0
requests>=2.31.0
psutil>=5.9.0

# Network and security
dnspython>=2.4.0
pycryptodome>=3.19.0
certifi>=2023.7.22

# Configuration and utilities
pyyaml>=6.0.1
click>=8.1.0
colorama>=0.4.6

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For building executables
# pyinstaller>=5.13.0
# nuitka>=1.8.0 